create chain ext lib, move examples to subdir
