+++ contracts/idn-contract-lib/examples/rand-extension-example/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 
 [dependencies]
 ink = { version = "5.1.1", default-features = false }
+idn-contract-lib = { path = "../idn-contract-lib", default-features = false }
 
 [lib]
 path = "lib.rs"
@@ -15,5 +16,6 @@ path = "lib.rs"
 default = ["std"]
 std = [
     "ink/std",
+    "idn-contract-lib/std",
 ]
 ink-as-dependency = []
