+++ contracts/idn-example-consumer-contract/lib.rs
@@ -569,7 +569,6 @@ mod example_consumer {
 		}
 	}
 
-	
 	#[cfg(all(test, feature = "e2e-tests"))]
 	mod e2e_tests {
 		use super::*;
